requirements.txt
# Python dependencies for Mini Cluster Setup

# Core configuration dependency (used by all containers)
PyYAML==6.0.1

# Data generator dependencies  
confluent-kafka==2.5.3
faker==25.9.2

# Spark containers automatically include PyYAML via custom Dockerfile

# Optional: For local development/testing
# pyspark==3.5.1
